Updating the English File
I took the liberty of cleaning up a lot of the text, while still preserving the "Lore" aspect of it. I'd like to do an overhaul to make the whole book feel like it was written "in-universe" by somebody who experienced these Ender items, then vanished, leaving just the book behind, but for now, this should clarify how the game mechanics work.